updating repo structure and moving slack manager class and config out of main file.
